-rename RT to RTStruct -move RTSS formatting out of contour to segmentation function -refactor generateROIContoursFromSegmentations to generateContourSetFromSegmentation -add check if no volume is found when generating contousr from segmentation -rename contourUtils to utilities -move helper functions out of RTSS.js to utilities folder -clean unnecessary console logs
