--- packages/adapters/src/adapters/Cornerstone3D/RTStruct/contourUtils/index.js
@@ -1,5 +1,0 @@
-import contourFinder from "./contourFinder";
-import mergePoints from "./mergePoints";
-import detectContourHoles from "./detectContourHoles";
-
-export { contourFinder, mergePoints, detectContourHoles };
