+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/contourFinder.js
