+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/detectContourHoles.js
