+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/index.js
@@ -1,0 +1,20 @@
+import contourFinder from "./contourFinder";
+import mergePoints from "./mergePoints";
+import detectContourHoles from "./detectContourHoles";
+
+import getPatientModule from "./getPatientModule";
+import getReferencedFrameOfReferenceSequence from "./getReferencedFrameOfReferenceSequence";
+import getReferencedSeriesSequence from "./getReferencedSeriesSequence";
+import getRTSeriesModule from "./getRTSeriesModule";
+import getStructureSetModule from "./getStructureSetModule";
+
+export {
+    contourFinder,
+    mergePoints,
+    detectContourHoles,
+    getPatientModule,
+    getReferencedFrameOfReferenceSequence,
+    getReferencedSeriesSequence,
+    getRTSeriesModule,
+    getStructureSetModule
+};
