+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/mergePoints.js
