+++ packages/adapters/src/adapters/Cornerstone3D/index.ts
@@ -13,7 +13,7 @@ import Length from "./Length";
 import PlanarFreehandROI from "./PlanarFreehandROI";
 import Probe from "./Probe";
 import * as Segmentation from "./Segmentation";
-import * as RT from "./RTStruct";
+import * as RTStruct from "./RTStruct";
 
 const Cornerstone3DSR = {
     Bidirectional,
@@ -33,7 +33,7 @@ const Cornerstone3DSR = {
 
 const Cornerstone3DSEG = {
     Segmentation,
-    RT
+    RTStruct
 };
 
 export { Cornerstone3DSR, Cornerstone3DSEG };
